Added command to create/update the mapping table

+++ src/CopyTheWorld/CopyTheWorld.Shared/DeviceMessages/DeskMotionSensorMetaData.cs
@@ -2,6 +2,6 @@
 {
     public sealed class DeskMotionSensorMetaData
     {
-        public int BatteryLevel { get; set; }
+        public double BatteryLevel { get; set; }
     }
 }
